requirements.txt
telethon>=1.41.2
python-dotenv==1.0.0
flask==2.3.2
python-telegram-bot==20.3
aiohttp==3.9.0
